+++ flexiapi/app/Http/Controllers/Api/Admin/VcardsStorageController.php
@@ -57,7 +57,7 @@ class VcardsStorageController extends Controller
         $vcard->vcard = preg_replace('/\r\n?/', "\n", $vcardo->serialize());
         $vcard->save();
 
-        return $vcard->vcard;
+        return $vcard;
     }
 
     public function update(Request $request, int $accountId, string $uuid)
@@ -76,7 +76,7 @@ class VcardsStorageController extends Controller
         $vcard->vcard = preg_replace('/\r\n?/', "\n", $vcardo->serialize());
         $vcard->save();
 
-        return $vcard->vcard;
+        return $vcard;
     }
 
     public function destroy(int $accountId, string $uuid)